+++ tidy3d/plugins/mode/mode_solver.py
@@ -179,6 +179,12 @@ class ModeSolver(Tidy3dBaseModel):
         """Axis normal to the mode plane."""
         return self.plane.size.index(0.0)
 
+    @staticmethod
+    def plane_center_tangential(plane) -> tuple[float, float]:
+        """Mode lane center in the tangential axes."""
+        _, plane_center = plane.pop_axis(plane.center, plane.size.index(0.0))
+        return plane_center
+
     @cached_property
     def solver_symmetry(self) -> Tuple[Symmetry, Symmetry]:
         """Get symmetry for solver for propagation along self.normal axis."""
@@ -723,6 +729,7 @@ class ModeSolver(Tidy3dBaseModel):
             symmetry=symmetry,
             direction=self.direction,
             precision=self._precision,
+            plane_center=self.plane_center_tangential(self.plane),
         )
 
         fields = self._postprocess_solver_fields(
@@ -778,6 +785,7 @@ class ModeSolver(Tidy3dBaseModel):
             direction=self.direction,
             solver_basis_fields=solver_basis_fields,
             precision=self._precision,
+            plane_center=self.plane_center_tangential(self.plane),
         )
 
         fields = self._postprocess_solver_fields(